// src/app/interfaces/personalInfo.obj.ts
interface PersonalInfoObj{
  id?: string;
  profile_image?: string;
  name?: string;
  apellidos?: string;
  description?: string;
  frase?: string;
  telefono?: number;
  correo?: string;
  competencias?: string;
  lenguajes_programacion?: string;
  frameworks?: string;
}

export {
  PersonalInfoObj
};